feat: add more endpoinst and error cases

+++ src/notes/notes.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Note } from './note.entity';
@@ -16,11 +16,31 @@ export class NotesService {
     return this.noteRepository.save(note);
   }
 
-  async findAll(): Promise<Note[]> {
+  async getAllNotes(): Promise<Note[]> {
     return this.noteRepository.find({ relations: ['project'] });
   }
 
-  async findOne(id: number): Promise<Note> {
+  async getNoteById(id: number): Promise<Note> {
+    const note = this.noteRepository.findOne({
+      where: { id },
+      relations: ['project'],
+    });
+
+    if (!note) {
+      throw new NotFoundException('Note not found');
+    }
+    return note;
+  }
+
+  async update(id: number, data: CreateNoteDto): Promise<Note> {
+    const note = await this.getNoteById(id);
+
+    if (!note) {
+      throw new NotFoundException('Note not found');
+    }
+
+    await this.noteRepository.update(id, data);
+
     return this.noteRepository.findOne({
       where: { id },
       relations: ['project'],
@@ -28,6 +48,20 @@ export class NotesService {
   }
 
   async remove(id: number): Promise<void> {
+    const note = await this.getNoteById(id);
+
+    if (!note) {
+      throw new NotFoundException('Note not found');
+    }
+
     await this.noteRepository.delete(id);
+  }
+
+  async getAllByProject(projectId: string): Promise<Note[]> {
+    return this.noteRepository
+      .createQueryBuilder('note')
+      .leftJoinAndSelect('note.project', 'project')
+      .where('project.id = :projectId', { projectId })
+      .getMany();
   }
 }
