+++ src/users/users.service.ts
@@ -1,4 +1,8 @@
-import { ConflictException, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
@@ -19,22 +23,40 @@ export class UsersService {
   }
 
   async getUserByEmail(email: string): Promise<User> {
-    return this.usersRepository.findOne({
+    const user = await this.usersRepository.findOne({
       where: { email },
       relations: this.relations,
     });
+
+    if (!user) {
+      throw new NotFoundException('user not found');
+    }
+
+    return user;
   }
 
   async createUser(user: UserDto): Promise<User> {
-    const isExistsUser = await this.getUserByEmail(user.email);
-    if (!!isExistsUser) {
+    const isExistsUser = await this.usersRepository.findOne({
+      where: { email: user.email },
+      relations: this.relations,
+    });
+
+    if (isExistsUser) {
       throw new ConflictException('user already exists');
     }
+
     return this.usersRepository.save(user);
   }
 
   async updateUser(email: string, data: Partial<User>): Promise<User> {
+    const user = await this.getUserByEmail(email);
+
+    if (!user) {
+      throw new NotFoundException('user not found');
+    }
+
     await this.usersRepository.update(email, data);
+
     return this.usersRepository.findOne({
       where: { email },
       relations: this.relations,
@@ -42,6 +64,12 @@ export class UsersService {
   }
 
   async deleteUser(email: string): Promise<void> {
+    const user = await this.getUserByEmail(email);
+
+    if (!user) {
+      throw new NotFoundException('user not found');
+    }
+
     await this.usersRepository.delete(email);
   }
 }
